refactor(codegen): consolidate parsers component schema to common
