+++ packages/react-native-codegen/src/parsers/typescript/components/index.js
@@ -12,7 +12,7 @@
 import type {ExtendsPropsShape} from '../../../CodegenSchema.js';
 import type {TypeDeclarationMap} from '../../utils';
 import type {CommandOptions} from './options';
-import type {ComponentSchemaBuilderConfig} from './schema.js';
+import type {ComponentSchemaBuilderConfig} from '../../schema.js';
 
 const {getTypes} = require('../utils');
 const {getCommands} = require('./commands');
